refactor: move file selection registry to frontend
