+++ packages/extension/src/frontend/files/fileSelectionRegistry.ts
