+++ packages/extension/src/frontend/files/index.ts
@@ -2,3 +2,10 @@
 export { FileLister, getFileLister } from './fileLister';
 export type { ListableFileType } from '@common/files/fileListingRules';
 export { getFilesRecursively } from './listing';
+export {
+  FILE_SELECTION_COMMAND_IDS,
+  MULTIPLE_FILE_COMMANDS,
+  type FileSelectionCommandId,
+  type FileSelectionResponseCommand,
+  type MultiFileCategory,
+} from './fileSelectionRegistry';
