+++ src/common/files/index.ts
@@ -11,10 +11,3 @@ export {
   type ListableFileType,
   type PreparedFileFilters,
 } from './fileListingRules';
-export {
-  FILE_SELECTION_COMMAND_IDS,
-  MULTIPLE_FILE_COMMANDS,
-  type FileSelectionCommandId,
-  type FileSelectionResponseCommand,
-  type MultiFileCategory,
-} from './fileSelectionRegistry';
